Refactor package structure to higherkindness.compendium
