+++ src/main/scala/higherkindness/compendium/db/DBServiceStorage.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 47 Degrees, LLC. <http://www.47deg.com>
+ * Copyright 2018-2019 47 Degrees, LLC. <http://www.47deg.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package higherkindness.db
+package higherkindness.compendium.db
 
 import cats.effect.Sync
 import cats.syntax.functor._
 import cats.syntax.flatMap._
-import higherkindness.models.Protocol
-import higherkindness.storage.Storage
+import higherkindness.compendium.models.Protocol
+import higherkindness.compendium.storage.Storage
 
 object DBServiceStorage {
