+++ src/main/scala/higherkindness/compendium/storage/Storage.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 47 Degrees, LLC. <http://www.47deg.com>
+ * Copyright 2018-2019 47 Degrees, LLC. <http://www.47deg.com>
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package higherkindness.storage
+package higherkindness.compendium.storage
 
-import higherkindness.models.Protocol
+import higherkindness.compendium.models.Protocol
 
 trait Storage[F[_]] {
